Let remainingFlow become 125% of realOptFlow, rather than raw optimal… (#841)
* Let remainingFlow become 125% of realOptFlow, rather than raw optimal flow.
This can fix the inconsistance between optimal flow in tooltip and that in turbine.

* criteria should be totalFlow != realOptFlow in HP Steam and Steam turbine

* fluidIntoPower should be long type, because ichorium huge turbine can generate power greater than Integer.MAX_VALUE on SC Turbine.

* Final change.

* 1 -> 1L

* revert the accidentally refactor in Plasma turbine

---------

+++ src/main/java/gtPlusPlus/xmod/gregtech/common/tileentities/machines/multi/production/turbines/GT_MTE_LargeTurbine_Gas.java
@@ -107,7 +107,7 @@ public class GT_MTE_LargeTurbine_Gas extends GregtechMetaTileEntity_LargerTurbin
     }
 
     @Override
-    int fluidIntoPower(ArrayList<FluidStack> aFluids, long aOptFlow, int aBaseEff, float[] flowMultipliers) {
+    long fluidIntoPower(ArrayList<FluidStack> aFluids, long aOptFlow, int aBaseEff, float[] flowMultipliers) {
         if (aFluids.size() >= 1) {
             int tEU = 0;
             int actualOptimalFlow = 0;
